Implement Sudoku board encoding and decoding functions
